Last update 21-12
- making a singletone - explain the threadssafe option and how it aborting creation of new instances: implemnt on
messegebroker, diary, squad, inventory, timeservice

in complete in messegebroker - do i need to add an option that the event is not exits?

in messegebroker methods of adding and broadcasting - do i need to create new user if it not exist in the list like guy work?

in the subscriber - sendbroad or event. is it threadsafe to check the condition ?



******
Semphore in squad - how to make it ???
timetick - how the system works?
in messegebroker -increase cuncourncy with syncronize only specific topics
unregister? and termination of all threads?